Enhance media player service with notifications and controls
- Refactored `MediaPlayerService` to use `MediaSessionCompat` for media control.
- Added notification support with media style and controls (play/pause, prev/next, stop).
- Implemented foreground service with dynamic notification updates for song progress.
- Added `resume` functionality to `BaseMediaPlayer` and `MediaPlayerImpl`.
- Integrated service control from `PlaySongFragment` using intents with actions (PLAY, PAUSE, RESUME, STOP, PREVIOUS, NEXT, SEEK_TO).
- Removed unused code and improved error handling in `MediaPlayerImpl`.
- Updated `Constants` with new `MediaAction` values and `POSITION` argument.
- Removed `MyApplication` notification channel setup as it's now handled in the service.
- Removed `MusicActivity`'s service binding and direct access, now using intents.

+++ app/src/main/java/com/lambao/mrbeast/domain/service/media_player/MediaPlayerCallBack.kt
@@ -4,5 +4,5 @@ interface MediaPlayerCallBack {
     fun onStart()
     fun onComplete()
     fun onError(e: Exception)
-    fun onPositionChanged(position: Int, duration: Int)
+    fun onPositionChanged(position: Long, duration: Long)
 }
